+++ database/migrations/2025_06_15_144559_create_rockets_table.php
@@ -22,12 +22,12 @@ return new class extends Migration
             $table->enum('environment_zone', ['vacuum', 'atmosphere']); // environment, but probably i need to redefine it
             $table->foreignId('celestial_body_id');
             $table->integer('required_delta_v')->default(1);
-            $table->decimal('twr')->default(1)->comment('Thrust-to-weight ratio');
-            $table->decimal('dry_mass')->default(0);
-            $table->decimal('wet_mass')->default(0);
-            $table->decimal('fuel_mass')->default(0);
+            $table->float('twr')->default(1)->comment('Thrust-to-weight ratio');
+            $table->float('dry_mass')->default(0);
+            $table->float('wet_mass')->default(0);
+            $table->float('fuel_mass')->default(0);
             $table->integer('cargo_mass')->default(1);
-            $table->decimal('trust')->default(0);
+            $table->float('trust')->default(0);
         });
     }
